requirements.txt
Cython==0.23.5
hidapi==0.7.99.post14
